changed successive service calls to a chain of pubs/subs. This makes it more modular for different sim scenarios.

+++ rosflight_sim/include/rosflight_sim/fixedwing_forces_and_moments.hpp
@@ -158,7 +158,7 @@ public:
    */
   geometry_msgs::msg::WrenchStamped update_forces_and_torques(rosflight_msgs::msg::SimState x,
                                                               geometry_msgs::msg::Vector3Stamped wind,
-                                                              std::array<int, 14> act_cmds) override;
+                                                              std::array<uint16_t, 14> act_cmds) override;
 };
 
 } // namespace rosflight_sim
