+++ rosflight_sim/src/fixedwing_forces_and_moments.cpp
@@ -657,7 +657,7 @@ rcl_interfaces::msg::SetParametersResult Fixedwing::parameters_callback(const st
 
 geometry_msgs::msg::WrenchStamped Fixedwing::update_forces_and_torques(rosflight_msgs::msg::SimState x,
                                                                        geometry_msgs::msg::Vector3Stamped wind,
-                                                                       std::array<int, 14> act_cmds)
+                                                                       std::array<uint16_t, 14> act_cmds)
 {
   delta_.a =  (act_cmds[this->get_parameter("aileron_channel").as_int()] - 1500.0) / 500.0;
   delta_.e = -(act_cmds[this->get_parameter("elevator_channel").as_int()] - 1500.0) / 500.0;
